MULE-19605: Nested elements should also have the concepto of min and max occurs
