+++ src/main/java/org/mule/runtime/api/meta/model/nested/NestableElementModel.java
@@ -11,6 +11,8 @@ import org.mule.runtime.api.meta.model.ComponentModel;
 import org.mule.runtime.api.meta.model.ComponentModelVisitor;
 import org.mule.runtime.api.meta.model.ComposableModel;
 
+import java.util.Optional;
+
 /**
  * A definition of an element that can be contained by a {@link ComposableModel}.
  *
@@ -22,6 +24,23 @@ import org.mule.runtime.api.meta.model.ComposableModel;
 @NoImplement
 @Deprecated
 public interface NestableElementModel extends ComponentModel {
+
+  /**
+   * Represents the minimum amount of times that this component can be used inside the owning one.
+   *
+   * @return An int greater or equal to zero
+   * @since 1.4.0
+   */
+  int getMinOccurs();
+
+  /**
+   * {@link Optional} value which represents the maximum amount of times that this component can be used inside the owning one.
+   * {@link Optional#empty()} means unbounded.
+   *
+   * @return If present, a number greater or equal than zero.
+   * @since 1.4.0
+   */
+  Optional<Integer> getMaxOccurs();
 
   /**
    * @return whether or not {@code this} element is required for its owner element
