+++ src/main/java/org/mule/runtime/api/meta/model/nested/NestedChainModel.java
@@ -12,6 +12,8 @@ import org.mule.runtime.api.meta.model.ComponentModel;
 /**
  * Represents a {@link NestableElementModel} that makes reference to a chain of {@link ComponentModel}s. This chain can have any
  * number of components of the allowed stereotypes.
+ * <p>
+ * For this type of nested element, both {@link #getMinOccurs()} and {@link #getMaxOccurs()} will never be greater than 1.
  *
  * @since 1.0
  */
